+++ test/acceptance/multilayer.js
@@ -1371,6 +1371,64 @@ suite('multilayer', function() {
       );
     });
 
+    // See http://github.com/CartoDB/Windshaft/issues/157
+    test("req2params is called only once for a multilayer post",
+    function(done) {
+
+      var layergroup =  {
+        version: '1.0.1',
+        layers: [
+           { options: {
+               sql: 'select cartodb_id, ST_Translate(the_geom, 50, 0) as the_geom from test_table limit 2',
+               cartocss: '#layer { marker-fill:red; marker-width:32; marker-allow-overlap:true; }', 
+               cartocss_version: '2.0.1',
+               interactivity: [ 'cartodb_id' ]
+             } },
+           { options: {
+               sql: 'select cartodb_id, ST_Translate(the_geom, -50, 0) as the_geom from test_table limit 2 offset 2',
+               cartocss: '#layer { marker-fill:blue; marker-allow-overlap:true; }', 
+               cartocss_version: '2.0.2',
+               interactivity: [ 'cartodb_id' ]
+             } }
+        ]
+      };
+
+      var expected_token; 
+      Step(
+        function do_post()
+        {
+          var next = this;
+          assert.response(server, {
+              url: '/database/windshaft_test/layergroup',
+              method: 'POST',
+              headers: {'Content-Type': 'application/json' },
+              data: JSON.stringify(layergroup)
+          }, {}, function(res, err) { next(err,res); });
+        },
+        function check_post(err, res) {
+          if ( err ) throw err;
+          assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
+          var parsedBody = JSON.parse(res.body);
+          expected_token = parsedBody.layergroupid;
+          assert.equal(server.req2params_calls, 1);
+          return null;
+        },
+        function finish(err) {
+          var errors = [];
+          if ( err ) errors.push('' + err);
+          redis_client.exists("map_cfg|" +  expected_token, function(err, exists) {
+              if ( err ) errors.push(err.message);
+              assert.ok(exists, "Missing expected token " + expected_token + " from redis");
+              redis_client.del("map_cfg|" +  expected_token, function(err) {
+                if ( err ) errors.push(err.message);
+                if ( errors.length ) done(new Error(errors));
+                else done(null);
+              });
+          });
+        }
+      );
+    });
+
 
 
     // TODO: check lifetime of layergroup!