+++ NEWS.md
@@ -1,5 +1,5 @@
 Version 0.19.0
-2014-MM-DD
+2014-02-19
 
  New features: