+++ NEWS.md
@@ -5,6 +5,7 @@ Version 0.19.1
 
  - Do not send 0 timings to statsd (fixes "Bad line: 0,ms" messages)
  - Do not call req2params more then once on MapConfig creation (#157) 
+ - Do not invalidate renderer caches when NO cache_buster is given (#158)
 
 Version 0.19.0
 2014-02-19