+++ NEWS.md
@@ -7,6 +7,7 @@ Version 0.19.1
  - Do not call req2params more then once on MapConfig creation (#157) 
  - Do not invalidate renderer caches when NO cache_buster is given (#158)
  - Call afterLayergroupCreate only after MapConfig verification (#159)
+ - Drop LRU cache for "seen" layergroups (#160)
 
  Bug fixes: