+++ NEWS.md
@@ -1,5 +1,5 @@
 Version 0.19.2
-2014-MM-DD
+2014-02-27
 
  Enhancements: