+++ NEWS.md
@@ -4,6 +4,8 @@ Version 0.19.3
  Enhancements:
 
  - Log full connection string on PostgreSQL connection error (#165)
+ - Prefix statsd resulting from hits of / and /version endpoints
+
 
 Version 0.19.2
 2014-02-27